+++ src/test/java/woowacrew/security/SecurityConfigTest.java
@@ -1,4 +1,4 @@
-package woowacrew.utils.configuration;
+package woowacrew.security;
 
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Test;
@@ -7,7 +7,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.web.reactive.server.WebTestClient;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class WebMvcConfigTest {
+class SecurityConfigTest {
 
     @Autowired
     private WebTestClient webTestClient;
